Removed duplicate folders in data structures
