+++ data_structures/Cpp/singly_linked_list.cpp
