+++ data_structures/Cpp/ternary_search.cpp
